+++ app/src/main/java/com/rootstrap/android/ui/viewmodel/ProfileViewModel.kt
@@ -1,14 +1,12 @@
-package com.rootstrap.android.ui.activity.main
+package com.rootstrap.android.ui.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.rootstrap.android.network.managers.session.SessionManager
 import com.rootstrap.android.network.managers.user.UserManager
-import com.rootstrap.android.network.models.User
 import com.rootstrap.android.ui.base.BaseViewModel
 import com.rootstrap.android.util.NetworkState
-import com.rootstrap.android.util.dispatcher.DispatcherProvider
 import com.rootstrap.android.util.extensions.ApiErrorType
 import com.rootstrap.android.util.extensions.ApiException
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -16,28 +14,23 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-open class SignInActivityViewModel @Inject constructor(
+open class ProfileViewModel @Inject constructor(
     private val sessionManager: SessionManager,
-    private val userManager: UserManager,
-    private val dispatcher: DispatcherProvider
+    private val userManager: UserManager
 ) : BaseViewModel() {
 
-    private val _state = MutableLiveData<SignInState>()
-    val state: LiveData<SignInState>
+    private val _state = MutableLiveData<ProfileState>()
+    val state: LiveData<ProfileState>
         get() = _state
 
-    fun signIn(user: User) {
+    fun signOut() {
         _networkState.value = NetworkState.loading
-        // Avoid using hardcoded dispatcher this way can be mocked later
-        viewModelScope.launch(dispatcher.io) {
-            val result = userManager.signIn(user = user)
+        viewModelScope.launch {
+            val result = userManager.signOut()
             if (result.isSuccess) {
-                result.getOrNull()?.value?.user?.let { user ->
-                    sessionManager.signIn(user)
-                }
-
                 _networkState.value = NetworkState.idle
-                _state.value = SignInState.signInSuccess
+                _state.value = ProfileState.signOutSuccess
+                sessionManager.signOut()
             } else {
                 handleError(result.exceptionOrNull())
             }
@@ -49,12 +42,13 @@ open class SignInActivityViewModel @Inject constructor(
             exception.message
         } else null
 
+        _networkState.value = NetworkState.idle
         _networkState.value = NetworkState.error
-        _state.value = SignInState.signInFailure
+        _state.value = ProfileState.signOutFailure
     }
 }
 
-enum class SignInState {
-    signInFailure,
-    signInSuccess
+enum class ProfileState {
+    signOutFailure,
+    signOutSuccess
 }
