+++ app/src/main/java/com/rootstrap/android/ui/viewmodel/SignUpViewModel.kt
@@ -1,12 +1,14 @@
-package com.rootstrap.android.ui.activity.main
+package com.rootstrap.android.ui.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.rootstrap.android.network.managers.session.SessionManager
 import com.rootstrap.android.network.managers.user.UserManager
+import com.rootstrap.android.network.models.User
 import com.rootstrap.android.ui.base.BaseViewModel
 import com.rootstrap.android.util.NetworkState
+import com.rootstrap.android.util.dispatcher.DispatcherProvider
 import com.rootstrap.android.util.extensions.ApiErrorType
 import com.rootstrap.android.util.extensions.ApiException
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -14,23 +16,29 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-open class ProfileActivityViewModel @Inject constructor(
+open class SignUpViewModel @Inject constructor(
     private val sessionManager: SessionManager,
-    private val userManager: UserManager
+    private val userManager: UserManager,
+    private val dispatcher: DispatcherProvider
 ) : BaseViewModel() {
 
-    private val _state = MutableLiveData<ProfileState>()
-    val state: LiveData<ProfileState>
+    private val _state = MutableLiveData<SignUpState>()
+    val state: LiveData<SignUpState>
         get() = _state
 
-    fun signOut() {
+    fun signUp(user: User) {
         _networkState.value = NetworkState.loading
-        viewModelScope.launch {
-            val result = userManager.signOut()
+        // Avoid using hardcoded dispatcher this way can be mocked later
+        viewModelScope.launch(dispatcher.io) {
+            val result = userManager.signUp(user = user)
+
             if (result.isSuccess) {
+                result.getOrNull()?.value?.user?.let { user ->
+                    sessionManager.signIn(user)
+                }
+
                 _networkState.value = NetworkState.idle
-                _state.value = ProfileState.signOutSuccess
-                sessionManager.signOut()
+                _state.value = SignUpState.signUpSuccess
             } else {
                 handleError(result.exceptionOrNull())
             }
@@ -42,13 +50,12 @@ open class ProfileActivityViewModel @Inject constructor(
             exception.message
         } else null
 
-        _networkState.value = NetworkState.idle
         _networkState.value = NetworkState.error
-        _state.value = ProfileState.signOutFailure
+        _state.value = SignUpState.signUpFailure
     }
 }
 
-enum class ProfileState {
-    signOutFailure,
-    signOutSuccess
+enum class SignUpState {
+    signUpFailure,
+    signUpSuccess
 }
